Remove global variables in favor of closures
In order for our `command_spy` to be in scope in both our test dummy
classes and our test examples we thought the only easy way would be to
make it global. RSpec however has the virtue of making heavy use of
blocks and closures. This change attempts to leverage Ruby's lexical
scoping to fight the war against global vars.

Dollar signs are gross anyway is this PHP or something?

Check these out:
* http://myronmars.to/n/dev-blog/2012/10/using-closures-to-avoid-leaking-state
* https://github.com/bestie/objective-ruby-style-guide